moved plugins to brain_bridge
