--- helios_sim/src/brain_bridge/autonomy/systems/mod.rs
@@ -1,9 +1,0 @@
-// helios_sim/src/simulation/plugins/autonomy/systems/mod.rs
-
-mod estimation;
-mod pipeline_tick;
-mod spawn;
-
-pub use estimation::update_odom_frames;
-pub use pipeline_tick::run_pipeline_tick;
-pub use spawn::{spawn_autonomy_pipeline, spawn_odom_frames};
