+++ helios_sim/src/plugins/sensors/gps.rs
@@ -3,7 +3,7 @@ use nalgebra::Vector3;
 use rand_distr::{Distribution, Normal};
 use std::time::Duration;
 
-use crate::brain_bridge::autonomy::components::{AutonomyPipelineComponent, SensorPublishChannel};
+use crate::brain_bridge::components::{AutonomyPipelineComponent, SensorPublishChannel};
 use crate::core::{app_state::SimulationSet, prng::SimulationRng, transforms::EnuVector};
 use crate::prelude::*;
