+++ helios_sim/src/plugins/sensors/raycasting.rs
@@ -3,7 +3,7 @@ use avian3d::prelude::{SpatialQuery, SpatialQueryFilter};
 use bevy::prelude::*;
 use std::time::Duration;
 
-use crate::brain_bridge::autonomy::components::{AutonomyPipelineComponent, SensorPublishChannel};
+use crate::brain_bridge::components::{AutonomyPipelineComponent, SensorPublishChannel};
 use crate::config::structs::{LidarConfig, SensorConfig};
 use crate::core::transforms::FluVector;
 use crate::core::{app_state::SimulationSet, prng::SimulationRng};
